installed jss and jsx
