+++ my-app/package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "react-jss": "^7.2.0",
+    "react-jsx": "^1.0.0",
     "react-scripts": "1.0.14"
   },
   "scripts": {
